+++ release_notes.rst
@@ -21,6 +21,7 @@ Features:
 - Replaced wishbone.module.process.humanlogformat with template module
 - Removed @-prefix from default event fields
 - Added wishbone.module.flow.queueselect
+- Added wishbone.module.flow.count
 
 Bugfixes: